Start the revival of NowSoundLib, dropping vestiges of the C++ demo app, moving to most current Windows SDK and v142 platform tools, and fixing precompiled header bleh.

+++ NowSoundLib/SpatialAudioProcessor.cpp
@@ -205,7 +205,7 @@ PluginInstanceIndex SpatialAudioProcessor::AddPluginInstance(PluginId pluginId, 
 
 int SpatialAudioProcessor::GetPluginInstanceCount()
 {
-    return _pluginInstances.size();
+    return (int)_pluginInstances.size();
 }
 
 NowSoundPluginInstanceInfo SpatialAudioProcessor::GetPluginInstanceInfo(PluginInstanceIndex index)
